refactor: standardiser PDF-visning på MultiPdfViewer-stacken
- Flytt EmbeddedPdf og MultiPdfViewer til src/common/pdf/
- Refaktorer PdfViewer til å bruke MultiPdfViewer internt
- Oppdater DokumentViewer med ny import-sti
- Slett duplikate filer fra sak-mappen
