+++ frontend/src/common/pdf/EmbeddedPdf.tsx
