+++ frontend/src/common/pdf/PdfViewer.tsx
@@ -1,7 +1,7 @@
-import {hentJournalpostMetaDataOptions} from "@generated/@tanstack/react-query.gen";
-import {Box, InlineMessage, Select} from "@navikt/ds-react";
-import {useSuspenseQuery} from "@tanstack/react-query";
-import {useEffect, useState} from "react";
+import { hentJournalpostMetaDataOptions } from "@generated/@tanstack/react-query.gen";
+import { Box, InlineMessage } from "@navikt/ds-react";
+import { useSuspenseQuery } from "@tanstack/react-query";
+import { MultiPdfViewer } from "~/common/pdf/MultiPdfViewer";
 import styles from "./PdfViewer.module.css";
 
 interface Props {
@@ -12,43 +12,19 @@ export function PdfViewer({ journalpostId }: Props) {
     if (!journalpostId) {
         return <InlineMessage status="warning">Det er ikke noe dokument å vise frem</InlineMessage>;
     }
-    return <PdfViewer2 journalpostId={journalpostId} />;
+    return <PdfViewerMedData journalpostId={journalpostId} />;
 }
 
-function PdfViewer2({ journalpostId }: { journalpostId: string }) {
+function PdfViewerMedData({ journalpostId }: { journalpostId: string }) {
     const { data: journalpost } = useSuspenseQuery(
         hentJournalpostMetaDataOptions({
-            path: {
-                journalpostId: journalpostId,
-            },
+            path: { journalpostId },
         }),
     );
-    const [dokId, setDokId] = useState<string | undefined>(journalpost.dokumenter?.at(0)?.dokumentInfoId);
-
-    useEffect(() => {
-        setDokId(journalpost.dokumenter?.at(0)?.dokumentInfoId);
-    }, [journalpost.dokumenter]);
-
-    if (!dokId) {
-        return <InlineMessage status="warning">Det er ikke noe dokument å vise frem</InlineMessage>;
-    }
 
     return (
         <Box className={styles.pdfViewer}>
-            <Select label="Velg dokument" value={dokId} hideLabel onChange={(e) => setDokId(e.target.value)}>
-                {(journalpost?.dokumenter || []).map((d, index) => (
-                    <option key={d.dokumentInfoId} value={d.dokumentInfoId}>
-                        Dokument {index + 1} av {(journalpost?.dokumenter || []).length} - {d.tittel}
-                    </option>
-                ))}
-            </Select>
-            <embed
-                src={`/api/journalpost/${encodeURIComponent(journalpostId)}/${encodeURIComponent(dokId)}`}
-                width="100%"
-                height="1200px"
-                type="application/pdf"
-                title="Embedded PDF Viewer"
-            />
+            <MultiPdfViewer journalPoster={[journalpost]} />
         </Box>
     );
 }
